Class Product created, all code refactored.
